tradX/coinview/requirements.txt
django~=2.2
python-binance
TA-Lib
numpy
backtrader